Edits, corrections, clarifications up to 16.9.2025
Added images and info about the promotion of AGI in prime time on TV in 2009, 2011. Corrections of a few other spotted typos or mistakes. Added the review of quote from "Man and Thinking Machine" by BgGPT. Reordered the LLMs conclusions after the visionary excerpt that was evaluated. ... Thorisson (not ssen once); "Stack Theory has its contributions" (not it contributions ...