Abstract, Epilog, footnotes, captions on 18.9.2025
Last: "CPUs ... CPUs... the AI Lab in 2001. Removed the second "CPU". LOL. Abstract, Epilog, footnotes on 18.9.2025 Improvement and extensions of the abstract and epilog, edits of several footnotes and more clarification in the epilog. Refined punch line analogy with Terminator 2 epilog - more succinct and closer to the original. "If a machine, a next-token predictor, learned the value of prodigy’s contributions – maybe humans can too". Fixed the cited title of the other book, which was simplified: *  “The first modern AI Strategy was published in 2003 by an 18-years-old and repeated and implemented by the whole world: The Bulgarian prophecies: How would I invest one million for the Greatest Benefit for the Development of my Country?”, ..
(the original is in Bulgarian).
Title of the 2009 interview: reverted to “I Will Create a Thinking Machine that Will Self-Complexify" (instead of the simplified “I Will Create a Thinking Machine that Self-Complexifies") in the Abstract.
Before that: Added an excerpt from the 2005 predictions about Machine Translation  - in Bulgarian, without translation - use MT, LOL.